style: signup page css
