+++ src/pages/user/SignUpAdmin.jsx
@@ -1,10 +1,6 @@
-import React, { useEffect, useState } from 'react';
-import { Input } from '../../components/Input';
-import { useNavigate } from 'react-router-dom';
-import { cookies } from '../../shared/cookies';
+import React, { useState } from 'react';
 import api from '../../axios/api';
-import useTrueHook from '../../hooks/useTrueHook';
-import Certification from './Certification';
+import { useNavigate } from 'react-router-dom';
 import {
   StBackground,
   StForm,
@@ -15,32 +11,47 @@ import {
   StTextInput,
 } from './UserStyled';
 import { StFont, StSmallFont } from '../Welcome/WelcomeStyled';
+import { Input } from '../../components/Input';
+import Certification from './Certification';
 
 function SignUpAdmin() {
-  // 가드
-  useTrueHook();
+  const [admin, setAdmin] = useState('');
+  const [password, setPassword] = useState('');
+  const [passwordCheck, setpasswordCheck] = useState('');
+  const [username, setUsername] = useState('');
+  const [companyName, setCompanyName] = useState('');
+  const [certification, setCertification] = useState('');
 
-  const [admin, setAdmin] = useState({
-    email: '',
-    password: '',
-    passwordCheck: '',
-    userName: '',
-    companyName: '',
-    certification: '',
-  });
-
-  const onChangeHandler = e => {
-    const { value, name } = e.target;
-    setAdmin(old => {
-      return { ...old, [name]: value };
-    });
-  };
+  const [passwordError, setPasswordError] = useState('');
+  const [passwordCheckError, setPasswordCheckError] = useState('');
 
   const navi = useNavigate();
 
-  // form태그 핸들러
-  const submitBtnHandler = async e => {
-    e.preventDefault();
+  const handlePasswordChange = event => {
+    const value = event.target.value;
+    setPassword(value);
+
+    if (value.length < 8) {
+      setPasswordError('비밀번호는 최소 8자 이상이어야 합니다.');
+    } else {
+      setPasswordError('');
+    }
+  };
+
+  const handlepasswordCheckChange = event => {
+    const value = event.target.value;
+    setpasswordCheck(value);
+
+    if (value !== password) {
+      setPasswordCheckError('비밀번호와 일치하지 않습니다.');
+    } else {
+      setPasswordCheckError('');
+    }
+  };
+
+  const handleSubmit = async event => {
+    event.preventDefault();
+
     try {
       const response = await api.post('/users/signup/admin', admin);
       alert(`${admin.userName}님 회원가입을 축하합니다.`);
@@ -55,139 +66,154 @@ function SignUpAdmin() {
   };
 
   return (
-    <>
-      <StBackground>
-        <StOverall height="1060px">
-          <div
-            style={{
-              marginTop: '80px',
-            }}
-          >
-            <StLoginForm onSubmit={submitBtnHandler} height="880px">
-              <StForm height="880px">
-                <StFormBox>
-                  <StFont align="start" fontSize="28px">
-                    관리자 회원가입
-                  </StFont>
+    <StBackground height='1000px'>
+      <StOverall>
+        <div
+          style={{
+            marginTop: '80px',
+          }}
+        >
+          <StLoginForm onSubmit={handleSubmit}
+          height='815px'>
+            <StForm>
+              <StFormBox>
+                <StFont align="start" fontSize="28px">
+                  관리자 회원가입
+                </StFont>
 
-                  <StTextInput height="118px">
-                    <Certification
-                      email={admin.email}
-                      onChange={onChangeHandler}
-                      admin={admin}
-                      setAdmin={setAdmin}
-                    />
-                  </StTextInput>
+                <StTextInput>
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    marginBottom="10px"
+                  >
+                    사용자 이름
+                  </StSmallFont>
+                  <Input
+                    type="text"
+                    value={username}
+                    onChange={event => setUsername(event.target.value)}
+                    placeholder="이름을 입력하세요."
+                    required
+                  />
+                </StTextInput>
 
-                  <StTextInput>
-                    <StSmallFont
-                      width
-                      align="start"
-                      fontSize="0.875rem"
-                      weight="700"
-                      marginBottom="10px"
-                    >
-                      인증번호
-                    </StSmallFont>
-                    <Input
-                      type="text"
-                      value={admin.certification || ''}
-                      onChange={onChangeHandler}
-                      name="certification"
-                      placeholder="인증번호를 입력하세요."
-                      required
-                    />
-                  </StTextInput>
+                <StTextInput height="118px">
+                  <Certification admin={admin} />
+                </StTextInput>
 
-                  <StTextInput>
-                    <StSmallFont
-                      width
-                      align="start"
-                      fontSize="0.875rem"
-                      weight="700"
-                      marginBottom="10px"
-                    >
-                      비밀번호
-                    </StSmallFont>
-                    <Input
-                      type="password"
-                      value={admin.password || ''}
-                      onChange={onChangeHandler}
-                      name="password"
-                      placeholder="영문, 숫자, 특수문자를 조합하여 입력하세요.(8~16자)"
-                      required
-                    />
-                  </StTextInput>
+                <StTextInput>
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    marginBottom="10px"
+                  >
+                    비밀번호
+                  </StSmallFont>
+                  <Input
+                    type="password"
+                    value={password}
+                    onChange={handlePasswordChange}
+                    placeholder="영문, 숫자, 특수문자를 조합하여 입력하세요.(8~16자)"
+                    required
+                    minlength="8"
+                    maxlength="16"
+                  />
+                </StTextInput>
+                {passwordError && (
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    color="red"
+                  >
+                    {passwordError}
+                  </StSmallFont>
+                )}
 
-                  <StTextInput>
-                    <StSmallFont
-                      width
-                      align="start"
-                      fontSize="0.875rem"
-                      weight="700"
-                      marginBottom="10px"
-                    >
-                      비밀번호 확인
-                    </StSmallFont>
-                    <Input
-                      type="password"
-                      value={admin.passwordCheck || ''}
-                      onChange={onChangeHandler}
-                      name="passwordCheck"
-                      placeholder="영문, 숫자, 특수문자를 조합하여 입력하세요.(8~16자)"
-                      required
-                    />
-                  </StTextInput>
+                <StTextInput>
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    marginBottom="10px"
+                  >
+                    비밀번호 확인
+                  </StSmallFont>
+                  <Input
+                    type="password"
+                    value={passwordCheck}
+                    onChange={handlepasswordCheckChange}
+                    placeholder="영문, 숫자, 특수문자를 조합하여 입력하세요.(8~16자)"
+                    required
+                    minlength="8"
+                    maxlength="16"
+                  />
+                </StTextInput>
+                {passwordCheckError && (
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    color="red"
+                  >
+                    {passwordCheckError}
+                  </StSmallFont>
+                )}
 
-                  <StTextInput>
-                    <StSmallFont
-                      width
-                      align="start"
-                      fontSize="0.875rem"
-                      weight="700"
-                      marginBottom="10px"
-                    >
-                      이름
-                    </StSmallFont>
-                    <Input
-                      type="text"
-                      value={admin.userName || ''}
-                      onChange={onChangeHandler}
-                      name="userName"
-                      placeholder="사용하실 이름을 입력하세요."
-                      required
-                    />
-                  </StTextInput>
+                <StTextInput>
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    marginBottom="10px"
+                  >
+                    회사
+                  </StSmallFont>
+                  <Input
+                    type="text"
+                    id="companyName"
+                    value={companyName}
+                    onChange={event => setCompanyName(event.target.value)}
+                    placeholder="회사를 입력하세요."
+                    required
+                  />
+                </StTextInput>
 
-                  <StTextInput>
-                    <StSmallFont
-                      width
-                      align="start"
-                      fontSize="0.875rem"
-                      weight="700"
-                      marginBottom="10px"
-                    >
-                      회사
-                    </StSmallFont>
-                    <Input
-                      type="text"
-                      value={admin.companyName || ''}
-                      onChange={onChangeHandler}
-                      name="companyName"
-                      placeholder="회사를 입력하세요."
-                      required
-                    />
-                  </StTextInput>
-
-                  <StLongButton type="submit">시작하기</StLongButton>
-                </StFormBox>
-              </StForm>
-            </StLoginForm>
-          </div>
-        </StOverall>
-      </StBackground>
-    </>
+                <StTextInput>
+                  <StSmallFont
+                    width
+                    align="start"
+                    fontSize="0.875rem"
+                    weight="700"
+                    marginBottom="10px"
+                  >
+                    인증번호
+                  </StSmallFont>
+                  <Input
+                    type="text"
+                    id="certification"
+                    value={certification}
+                    onChange={event => setCertification(event.target.value)}
+                    placeholder="인증번호를 입력하세요."
+                    required
+                  />
+                </StTextInput>
+                <StLongButton type="submit">Submit</StLongButton>
+              </StFormBox>
+            </StForm>
+          </StLoginForm>
+        </div>
+      </StOverall>
+    </StBackground>
   );
 }
-
 export default SignUpAdmin;
